+++ CHANGELOG.md
@@ -6,6 +6,20 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+### Changed
+
+### Deprecated
+
+### Removed
+
+### Fixed
+
+### Security
+
 ## [0.2.1] - 2021-01-21
 
 ### Added
@@ -129,6 +143,8 @@ and this project adheres to
 
 - Initial release.
 
+[unreleased]:
+  https://github.com/daxpedda/sentry-contrib-native/compare/0.2.1...HEAD
 [0.2.1]: https://github.com/daxpedda/sentry-contrib-native/releases/tag/0.2.1
 [0.2.0]: https://github.com/daxpedda/sentry-contrib-native/releases/tag/0.2.0
 [0.1.0]: https://github.com/daxpedda/sentry-contrib-native/releases/tag/0.1.0