+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to
 ### Changed
 
 - Changed `zlib` for Crashpad to always build from source.
+- Updated `sentry-native` to `c95dbb0`.
 
 ### Deprecated